WIP: Updates for next `calloop` version, using `BorrowedFd`
Supersedes https://github.com/Smithay/smithay/pull/839, the other parts
of which should be merged now.

This is blocked on release of `calloop` and `udev`.

The same awkwardness exists as in the previous PR around
`display.backend().poll_fd()`. That returns a `BorrowedFd<'_>` that we
can't simply insert into an event loop.

This is also a bit awkward with `child_stdout` in `xserver.rs` (I guess
we want to remove the event source without closing that?)