WIP winit git
I wanted to see how the changes in Winit git would work here. So of
course this depends on that being released.

Depends on https://github.com/Smithay/smithay/pull/1514. It doesn't seem
to pose a particular challenge beyond that. Updating compositors for
that change will be needed to test this...